requirements.txt
fastapi==0.109.0
uvicorn==0.29.0
openai==1.3.7
pgvector==0.2.2
sentence-transformers==2.2.2
python-dotenv==1.0.0
langgraph==0.1.1
azure-search-documents==11.4.0
requests==2.31.0
azure-storage-blob==12.16.0